cmake: remove GCC-9 -Wno-error flags added d/t "Qt not ready"
We recently integrated 12313f3ebe00808e9b1c27ddcb49cedf3e09582b
containing a bug that would have been caught by the CI if it weren't
for a -Wno-error=implicit-fallthrough. Like for fallthroughs, I
haven't seen these warnings in a long time in Qt, so turn them back
into errors.

This episode should show that the -Wno-error approach is wrong, unless
we have a clear time limit for lifting it again. If these flags are
allowed to linger, mistakes are caught later.

Pick-to: 6.10
Change-Id: I97ebc9d17ee97eb5d1d35506b11f0daea7e50c9d
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>