a11y: Set buddies in QFileDialog
This allows quickly navigating to the combo boxes
labelled by the "Look in:" and "Files of type:" labels
and makes sure the accessible relations between the label
and the comboboxes are set.
This makes screen readers like Orca on Linux announce the
label texts when the comboboxes receive focus, so the
meaning of the comboboxes becomes clear.

Update tst_QFiledialog::labelText to include the
newly set shortcut keys in the expected label texts.

Fixes: QTBUG-141641
Change-Id: I8bc156d5e8d9b337a2a7ff34fe93ff1dda393de9
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>