QColorPicker: Move signal emission to QColorPicker::setCol
Instead of emitting the signal in both callers of the
QColorPicker::setCol variant taking a QPoint, emit
the signal from that method.

This also prepares for further uses of the method
in an upcoming commit without having to take care of
manually emitting the signal in those as well.

Task-number: QTBUG-141666
Change-Id: I2a373ebdb983e458d8393effc13bc4abf582fb25
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>