QFileSystemModel: rmdir() should call removeNode()
This fixes a memory leak in
tst_QFileSystemModel::pathWithTrailingSpace().

Creating a dir with QFSM::mkdir() creates and inserts a node in the
`children` QHash of its parent node. Removing the dir didn't
remove/delete the created node (it would have eventually been deleted
by the qDeleteAll() in the dtor of its parent node). However, we rename
another dir to the same name, overwriting the node in the hash without
deleting it.

Found with ASAN.

Direct leak of 720 byte(s) in 9 object(s) allocated from:
    #0 0x7fdfa2f2273b in operator new(unsigned long) (/lib64/libasan.so.8+0x12273b) (BuildId: cbfe49f3b7600c4f194d4c54774c977296e9d98a)
    #1 0x7fdf870672fc in QFileSystemModelPrivate::addNode(QFileSystemModelPrivate::QFileSystemNode*, QString const&, QFileInfo const&) src/gui/itemmodels/qfilesystemmodel.cpp:1855
    #2 0x7fdf8705e2f0 in QFileSystemModel::mkdir(QModelIndex const&, QString const&) src/gui/itemmodels/qfilesystemmodel.cpp:1468
    #3 0x5600812751f1 in tst_QFileSystemModel::pathWithTrailingSpace() tests/auto/gui/itemmodels/qfilesystemmodel/tst_qfilesystemmodel.cpp:1359

Amends the beginning of the repo.

Fixes: QTBUG-135038
Pick-to: 6.10 6.8 6.5
Change-Id: I9849784efacf8eac9541e8bd5d78bbd9d0b771cb
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>