Make QTZP::findLongNamePrefix() cope with zones unknown to CLDR
The backend may know about zones for which our most recent CLDR update
lacked data, for example because one has been added to the IANA
database recently, as has happened for America/Coyhaique in release
2025b. Because findLongNamePrefix() was working with a local sui
generis type to hold details by which to look up an IANA ID in the
CLDR tables, if it found a zone that matched for the locale's region
format, it had to find the index of that zone in the tables to fill in
the sui generis type's field, even though that was only used to look
up the IANA ID later. That lookup failed if the IANA ID wasn't in the
CLDR tables. So switch to using a NamePrefixMatch to hold details in
the final region format search, so that we can simply store the ID
rather than an index that might not exist.

Add the new zone that triggered this to test data and fix alphabetic
order in its list while I'm about it.

Amends commit 2edd9286cf386675be76032424248e60216f6331 (6.10).

Pick-to: 6.10
Change-Id: I7c4850f8d368460852c7d19801ee80cf25ccf7e8
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>