QColorPicker: Remember position, not hue and sat
Instead of remembering the hue and saturation value
and calculating those when setting a new position
by using the mouse, instead do the opposite:
Remember the position and calculate the hue and
saturation values on demand.

The motivation is that translating a position to a hue/sat
value and then translating that back to a position
doesn't necessarily restore the exact original position,
but may result in a position that slightly differs
(e.g. shifted by a pixel, because multiple pixel
positions can represent the same hue/saturation
values).

While that's not a big deal when it results in the position
getting set to one that's a pixel away from the exact
position that was clicked using the mouse, it would become
a problem with an upcoming commit that allows to move the
position using the keyboard:
Pressing/Holding an arrow key should result in the
position moving in the expected direction, but that
wouldn't be the case with the previous logic.
(In a first attempt without this commit in place,
it would result in the position moving more or less
diagonally towards the top-left for the Left and Up
keys, upwards for the Right arrow key and left for
the Down key.)

In the ctor, call base class method QWidget::adjustSize
to ensure that the widget has its final (fixed) size,
which allows creating the pixmap there.
However, still adjust QColorPicker::resizeEvent
to calculate the new position based on the
hue/saturation values that can be calculated
using the current position and previous widget size.
(In a quick test using GammaRay [1] to manually
change the widget size, this results in the position
getting updated as expected, i.e. the cross is then
shown at the new position of the currently selected
color.)

Make sure to keep the hue and saturation values
within the valid range in QColorPicker::huePt and
QColorPicker::satPt.
Previously, QColorPicker::setCol(int, int) would
take care of it, but QColorPicker::setCol(const QPoint &)
no longer calls this method now, but now the former
calls the latter. Therefore, move some logic from the
former to the latter so it still gets run in both cases.

Keep previous logic to not emit the QColorPicker::newCol
signal when the QColorPicker::setCol(int, int) setter
gets called as the QColorDialog implementation relies
on that.

[1] https://github.com/KDAB/GammaRay

Task-number: QTBUG-141666
Change-Id: I72d9cfd8920e4b608baa07c5728ee358eee55823
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>