androidtestrunner: do not exit(-1) when things go sideways
Negative exitcodes are platform-dependent, in fact on Linux exit(-1) is
received as 255 from the parent process.

We avoid exitcodes [0,127] because QTest-based tests will, under normal
circumstances, return values within [0,127] and androidtestrunner's job
is to return that exact exitcode.

We also avoid 255 itself to avoid confusion with the previous version.

Finally we avoid exit codes close to 128 as some parent processes use
those to signify termination by signal (e.g. Ctrl-C of a process under
bash usually shows $? as 130).

Therefore, we define EXIT_ERROR to be 254 instead of -1. It signifies
some generic, uncategorized failure in androidtestrunner.

We also define some more exit codes 251-253 for more specific things
that went wrong.

Finally we change relevant output from androidtestrunner to be prefixed
by "[androidtestrunner] " so that important messages are not lost in the
big amount of output that android builds spew. Ideally all lines would
be prefixed, but that can come later with a wrapper around
qDebug(), qCritical() etc.

Pick-to: 6.11 6.10 6.8
Change-Id: I63edb9a429029ea7c1c9caa35fd7521832e3f9a8
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>