macOS: NFC normalize selected file in QNSOpenSavePanelDelegate
Apple systems with HFS+ were traditionally storing file names in NFD
normalized form, while Qt prefers to operate in NFC, as that is more
likely to match what the user types with the keyboard.

Nowadays APFS is truly normalization preserving, in that it stores
the filename in whatever normalization the filename was written as,
but this still means we might encounter filenames in NFD form, so
we keep normalizing file paths to ensure Qt and user code can do
file name comparisons without needing to deal with potentially
different normalizations.

Pick-to: 6.10 6.8 6.5
Task-number: QTBUG-122624
Change-Id: Iaed71c8194b44c82b21a0bb3ff34fdd81c8325bd
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>