Don't emit QFileDialog::directoryEntered when not changing dir
In QFileDialogPrivate::enterDirectory, return early
if the directory to enter is already the current
directory.

This prevents emitting the QFileDialog::directoryEntered
signal when no new directory was actually entered.

Extend tst_QFiledialog::directoryEnteredSignal
accordingly.

This prepares for an upcoming fix for QTBUG-141644
where the fact that QFileDialogPrivate::enterDirectory
was called again with the same path resulted in
the signal getting emitted twice and therefore
causing the existing

    QCOMPARE(spyDirectoryEntered.size(), 1);

check in tst_QFiledialog::directoryEnteredSignal to fail.

Task-number: QTBUG-141644
Change-Id: I3359f9f8af3ce03c7b32c040a92d4bf09e065e85
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>