a11y: Let QAccessibleTable subclass QAccessibleWidgetV2
Since QAccessibleTable and subclasses are a11y implementations
for item views and those are widgets, subclass
QAccessibleWidgetV2 instead of only the more generic
QAccessibleObject (which is itself a QAccessibleWidgetV2
base class).

This e.g. ensures that all accessibility properties set
on the QWidget (like a manually set accessible name
or description) or the QAccessible::Labelled relation
to a QLabel for which the widget was set as the buddy
are set correctly.

Task-number: QTBUG-141703
Change-Id: I7f5f1e6d3f84c97e541517baada637d533fa3e00
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>