a11y QFontDialog: Consistently set lists as buddies
As is done for the "Font" and "Font style" labels,
also set the corresponding list as the buddy of the
"Size" label and not the line edit.

This ensures consistent behavior and that the
label gets announced by screen readers when using the Tab
key to navigate through the dialog and the "Size" list view
receives focus.

Fixes: QTBUG-141703
Change-Id: Idb228a665e2f73c13516a8b4c1ebb63ba6d28b25
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>