Android: expect testFullScreenDimensions on Android 16 to fail in CI
This patch changes expectation for one condition in CI to fail.
Test or implementation needs later to be fixed, or test to be
removed. Also the edge-to-edge support should be changed later
to work in similarly at locally and at CI.

Change is only done to dev, as similar fail has not seen in 6.10.

Task-number: QTBUG-141712
Change-Id: I42452bdda0baa2b32d83b6f858b96cadb6c2f52e
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Ville Voutilainen <ville.voutilainen@qt.io>