Http: Set error to 'TimeoutError' for a transfer timeout
When transferTimeout was added for QNAM it was implemented
in terms of calling abort(), which is fine, it does what it
needs to do. But it sets OperationCanceledError, which is correct but
misleading since it's the same as for a user-initiated abort.

[ChangeLog][QtNetwork][QNetworkAccessManager] The error set when a
request times out due to QNetworkRequest::setTransferTimeout() has been
changed to QNetworkReply::TimeoutError, to disambiguate from the
previously used QNetworkReply::OperationCanceledError.

Pick-to: 6.10
Fixes: QTBUG-96165
Change-Id: Ibb2f28637f034190fb125630f96aa004c3634fef
Reviewed-by: Mate Barany <mate.barany@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>