macOS: Disable window-modal native message boxes on Tahoe
The combination of a window-modal NSAlert, and the message-box being
shown via exec() instead of open(), which results in us spinning a
nested [NSApp run], causes the dialog buttons to stop responding to
mouse events (while keyboard navigation works fine).

As this problem isn't seen for any other dialogs shown as sheets,
and the NSAlert still generally responds to mouse events (such as
text selection), this problem seems limited to NSAlert, so we
disable the native message box in this specific case.

[ChangeLog][macOS] Window-modal message boxes are no longer backed
by native NSAlerts on Tahoe, due to a bug that prevents the alert
from responding to mouse events for its buttons.

Fixes: QTBUG-141689
Pick-to: 6.10 6.8 6.5
Change-Id: I6fb24b84ce3d6fb14cfbce4a38079b67626239a8
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>