wayland: Fix regression with custom popup completers
Commit 4eed22e67ed98ec2565be84a4838955af0faf8b0 has broken the handling
of popup-based QCompleter that use custom positioning, e.g. those
attached to multi-line QTextEdit widgets - as in the "customcompleter"
example. Following that change, popups now appear above or below the
parent widget, which is usually far away from the cursor position or
whatever other "hot" rectangle the application needs. Treating the
completer popup the same as a combo dropdown is just not correct.

There is probably something better to do to allow the compositor to
place such popups correctly, but for now the best course of action
is to not opt-in to the new Wayland mechanism to avoid breakage.

Change-Id: I237d5f49b185c6a8593ef2dd72c9bcb6c6ff4785
Reviewed-by: David Edmundson <davidedmundson@kde.org>