Fix QRandomAccessAsyncFile::close() logic on macOS
The QRandomAccessAsyncFilePrivate::close() function is supposed to
cancel all operations and close all the related IO channels,
including the main IO channel.

The initial implementation had a false assumption that the callback
indicating that the main IO channel is closed is always the last
callback that is called in this scenario. So, that callback was
used to unconditionally call wakeOne() on the condition variable
that was used to signal that all operations are completed.

However, in practice when we cancel the operations, we receive the
following callbacks:
* close callback from each IO channel related to each operation;
* completion callback for read/write operations with done set to true
  and error code indicating cancellation;
* close callback from the main IO channel.
All these callbacks can come in an arbitrary order.

As a result, the case when multiple operations are still in progress,
and the file is removed (or close() is called) was randomly crashing
on CI, because the completion callbacks were trying to access the
already-removed file.

This patch updates the logic to only wake a condition variable when
all callbacks are received. In order to do that we introduce a new
m_numChannelsToClose variable to keep track of the number of active
channels, and use the size of the m_runningOps set to track the
running operations.

To verify that the approach works reliably, I manually ran the
equivalent of tst_QRandomAccessAsyncFile::fileRemovedInProgress()
100'000 times for each operation (owning and non-owning read/write
separately).

Amends e6b26ad1eca555ee1e6b2c4e9b4c7676dc754b04.

Fixes: QTBUG-141730
Change-Id: I874b9b9999398f787b5039c78bc3d3eee44c088c
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>