QVariant: remove pointless cast
Placement new wants to be called with a void* (cf. VOIDIFY() in
std::construct_at()'s definition), so stop there, don't continue to
cast to char*.

This was already done for the primary overload of the in-place
constructor, but forgotten in the initializer_list one.

Amends c81e8f8ff24d30cc4c137d6f071954958e984ce9.

Pick-to: 6.10 6.8
Change-Id: I5fe43d0062f2ad5e88603d5d231c2053f1164fed
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>