QStyleSheetStyle: don't call base implementation for PE_Widget
Don't call the base implementation for PE_Widget for an
QAbstractScrollArea when we have rules for it to avoid overriding of the
style paintings by the base class.
This was not visible until the windows11 style reimplemented the drawing
for PE_Widget with 072bf40fe9597bdd910bb0c1e6a754c1a468b844.

Pick-to: 6.10 6.8
Fixes: QTBUG-141819
Change-Id: I6e90dc6cb3b1a7151815d57d790c67317a97d589
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>