QRandomGenerator: remove direct use of HW instructions
The HWRNG instructions are not necessarily faster than what we can get
from the OS. Moreover, the OS is able to collect entropy from sources
other than the CPU itself. More importantly, this removes the need for
Qt to deal with broken HWRNG, passing the buck to the OS (which may
disable the CPUID bit, causing the application to fail to load).

[ChangeLog][QtCore][QRandomGenerator] This class no longer directly uses
a hardware random number generator on x86 systems, even if one is
available. Instead, it will always use a generator provided by the OS
(so performance will be OS-specific), though there should be no
meaningful difference in the quality of the samples generated.

Task-number: QTBUG-69423
Task-number: QTBUG-129193
Pick-to: 6.10 6.8 6.5
Change-Id: I5121c5a34d684983fa1dfffdbabd22de51966650
Reviewed-by: Jason A. Donenfeld <Jason@zx2c4.com>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>