QThread/Unix: intercept exceptions only with Glibc
See 5677b70eee2e923eea8e5150500ac745d8d54974 for details why we need
such specific exception handling for Glibc. In short, on Glibc,
pthread_cancel and pthread_exit are implemented by throwing a special
kind of exception that can be caught, but must always be rethrown. That
exception is then used to activate the cancellation clean-up handlers.

But for libc++ on Apple platforms we can simply wrap it as noexcept
to guarantee that std::terminate() will be called in case of any
unhandled exception.

Task-number: QTBUG-141803
Pick-to: 6.10 6.8
Change-Id: Iaa88d3a8091425206ee2735e835ae74fd087e9e0
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>