Do some minor tidy-up in qtimezonelocale.cpp
Its OffsetFormulaMatch::operator bool() was missing
const-qualification; supply it and use it.
Combine an indexOf() and the check of its return into an if condition.
Split up an if-else chain where an early return made else redundant.
Add notes with links to help clarify what's going on with some of the
formatting details.

Pick-to: 6.10 6.8
Task-number: QTBUG-139223
Change-Id: Ie235641c3086965aeb631588c55ec38a554a08e0
Reviewed-by: Mate Barany <mate.barany@qt.io>