Modernize tst_LargeFile: use NSDMI for construction where possible
The QEMU special case only affected the QT_LARGEFILE_SUPPORT branch of
the cascade of #if-ery on maxSizeBits, which can thus be constexpr in
all other cases and const even in that case. Also record why it's
limited to 28 rather than the 31 implied by the prior comment
(previously only otherwise discoverable in git history).

Mediate the complex #if-ery for the initial value of maxSizeBits by
introducing a constexpr maxAllowedSizeBits set within such #if-ery, so
as to simplify setting maxSizeBits itself. Note that this member is in
fact varied by some tests, implying order-dependence. Move the member
constants to before the constructor to make it possible to have a
define that correctly controls inclusion of the QEMU tweak.

Change-Id: Id6f8d55a34431d8f08f0e2b346c826bb49991423
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>