iOS: Import documents when opening files via UIDocumentPickerController
As part of implementing support for native folder dialogs in
dbe433f6d0259119fa31103b3428d50cc9fec286 we removed the import
mode logic, switching away from UIDocumentPickerModeImport for
files to effectively always preferring UIDocumentPickerModeOpen.

This regressed file opening, as unless the application explicitly
manages access to security scoped resources (files outside of its
sandbox) the files selected in the file dialog will be inaccessible.

We now pass asCopy:YES when we know we're opening files, which
restores the previous behavior.

For opening directories outside of our sandbox we can't pass
the same flag, as folders can't be imported. As a result, the
app can't see any content in the security scoped folders.

This will be improve in a follow up that adds support for
managing security scoped resources.

Pick-to: 6.10 6.10.1 6.8 6.5
Fixes: QTBUG-117832
Change-Id: Ifdc3866c065e777357a2d400f980c27efb8eb9af
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>