QThread/Unix: deduplicate conditionals for setCurrentThreadName()
Refactor the setCurrentThreadName() calls to avoid duplicating the
conditions under which the functionality is present. That is, do it all
inside of the function, and pass the necessary information into it.

For OSes that don't support any of the calls, this may call
.toLocal8Bit() unnecessarily.

Amends commit 3ef51efbe75bfb9f1dfbe7df073e9eb745a72ad8.

Pick-to: 6.11 6.10
Change-Id: I6401c8a9cc4e413dadf2fffd79f31eb63470a038
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>