QRM: report the number of properties as column count in trees of gadgets
Instead of hardcoding the column count to 1 for one-dimensional ranges,
use the row traits's fixed_size attribute, which reports the number of
properties for rows that are gadgets.

This was already correctly done for tables. To remove duplication, move
the structure-independent part of the logic into a reusable helper in
the base class.

Add basic test coverage.

Fixes: QTBUG-141885
Pick-to: 6.10
Change-Id: Ic10a8a61998184485acb89ca9db0d9c9be822f2d
Reviewed-by: David Faure <david.faure@kdab.com>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>