configure: Simplify handling of the -make and -nomake options
Remove the options '-make libs'. This was doing nothing.
We don't have a QT_BUILD_LIBS variable.

Remove the option '-make tools'. This was doing nothing. We could map
this to QT_FORCE_BUILD_TOOLS=ON, but this would be confusing for desktop
builds, because QT_FORCE_BUILD_TOOLS=ON yields an error if QT_HOST_TOOLS
is not set.

Remove the option '-nomake tools'. This was printing a "not yet
implemented warning" without any further effect. This could be mapped to
QT_FORCE_BUILD_TOOLS=OFF but people expect different behavior from this
option (e.g. disabling all apps).

The list of allowed values for -make and -nomake is now in one location:
qt_cmdline.cmake, and the superfluous handling of unknown input has been
removed from QtProcessConfigureArgs.cmake. Every value of -make /
-nomake now maps simply to QT_BUILD_<part>=ON/OFF.

Change-Id: I251dd6185d6ae1f22c90b2b8443ba008a6da01f1
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>