iOS: Implement support for native save dialogs
The iOS equivalent is a UIDocumentPickerController that exports an
existing file, so we need to create an empty temporary dummy file
to be able to show the dialog. This is also how we give a suggested
file name.

Once the target file is chosen, it's available through the normal
selected-file APIs, and the application can write to it through
the normal security scoped file system backend.

For some reason setting directoryURL to an inaccessible path,
such as the .app bundle, results in the document picker showing
an empty sheet. And in testing both open and save dialogs with
the directoryURL set to a known good path (Documents), the
iOS dialog still prefers to open to the user's last used path,
so we skip the entire logic for save dialogs for now, to be on
the safe side.

[ChangeLog][iOS] Native file dialogs on iOS now support saving
files.

Task-number: QTBUG-120528
Change-Id: I10a0e6ce78b23163933ecf67269c54b618c30964
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Doris Verria <doris.verria@qt.io>