QThread/Unix: re-enable cancellation before the event dispatcher
That makes it possible for the user-supplied dispatcher to abort a
thread start (however unlikely such an event might be).

[ChangeLog][Important Behavior Changes] POSIX Thread cancellations are
now enabled during the event dispatcher's startingUp() call and when the
QThread::started() signal is emitted. This allows the thread to be
terminated (e.g. via QThread::terminate()) during those calls.

Pick-to: 6.11
Change-Id: Ia4c60118f1eeefde3127fffd8d9d0a06968cc0e7
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>