QThread/Unix: update termination on exceptions leaving run() & co
Amends the original commit 5677b70eee2e923eea8e5150500ac745d8d54974
("QThread: terminate on exceptions leaving run()", Qt 5.10) without
changing behavior. Our code matched what libstdc++ used to be; this
commit matches the change in [1] which explains as "Remove try-block
so that exceptions propagate out of the thread and terminate is called
by the exception-handling runtime". This removes the now-unnecessary
terminate_on_exception() function added by commit
f0ffe351492a7d8e9fac8a8e060b906b4b8a733a.

The explanation is in the comment in the file: we don't need to
explicitly call std::terminate() to achieve this, because the lack of
an exception handler in our functions and in our callers will cause
the runtime to call it. With the added benefit of having the SIGABRT
backtrace pointing to the throw location, instead of the explicit call
to std::terminate() in our code.

[1] https://gcc.gnu.org/git/?p=gcc.git;a=commit;h=754d67d5ba4a1f9994210d402893a4cf49ce6a71

Fixes: QTBUG-141803
Pick-to: 6.11
Change-Id: Idb5528929d59b041aa52fffd45746868a25206dd
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>