Turn QHashPrivate::isRelocatable into a variable template
(was: function template)

Coverity complained that, when instantiating a QHash with
is_same<mapped_type, key_type>, the two terms of the logical AND
operator were identical.

This is, of course, expected and perfectly as intended, but in the
interest of fixing-instead-of-dismissing, try whether making the
entity a variable template intead of a function template will fix the
Coverity complaint, too. After all, this is how you would write the
entity today, in C++17.

As a variable template, add the idiomatic _v suffix (to explicitly
break out-of-tree users, if any, with a clear error message).

Amends 5b7c3e31b538376f2b4733bd868b5875b504cdb3.

Coverity-Id: 378449
Change-Id: I1dab5d18d6f55edd58e16b9773403cee2f93dfee
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Giuseppe D'Angelo <giuseppe.dangelo@kdab.com>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>