QHash: fix noexcept on Node::takeValue()
In the QHashDummyValue specialization, the noexcept was missing. Add
it.

In the primary Node template, takeValue() was conditionally noexcept,
but seeing as it just returns an rvalue reference to the 'value' data
member, it's of course unconditionally noexcept, since the actual move
contructor or -assignment operator call happens after the function
returns.

Besides, even if it had to be conditionally noexcept (e.g., because it
returned T-by-value), it would have to check for
nothrow_move_constructible, not _assignable.

Amends 5b7c3e31b538376f2b4733bd868b5875b504cdb3.

Picking to older branches is ok, since these functions, even if they
were marked noexcept(false), could not throw any exceptions, so
forward BC is not broken.

Pick-to: 6.10 6.8 6.5
Change-Id: I43aab284acc8b1303d4147c1b8c58613bbb05b25
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>