CMake: Skip creating feature header files if NO_HEADERS is specified
If NO_HEADERS option is specified to qt_feature_module_begin(),
skip generating the public and private qfoo_config.h and
qfoo_config_p.h files.

We can't tie this behavior to the NO_MODULE option, because in the
qtbase GlobalConfig case, we specify NO_MODULE but still expect
the headers to be created.

Also allow not specifying the PUBLIC_FILE and PRIVATE_FILE arguments
when NO_HEADERS is specified.

Pick-to: 6.8 6.10
Change-Id: I006f35d269910d13899827d4149d4c060c3d006a
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>