QIORing: fast-track cancellation
It only makes sense for a full queue, which we don't really test with,
because in a normal case we are preparing the task immediately, so it
will be queued already, meaning we can't benefit from it being
'unqueued'. See follow-up for the tests.

We could, of course, start preparing tasks at the same time as pushing
them to the kernel, but I'd rather not, in case an immediate failure
shows up.

Pick-to: 6.11
Change-Id: Ib445d5ebfabbbaf01e65eac4152cf1bd58ea6041
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>