QIORing/Linux: handle >2gib reads&writes
Linux internally truncates any read and write at a limit of
'MAX_RW_COUNT', which is slightly less than 2 GiB, for a few reasons.

We will work around it by reading or writing some segments at a time,
re-issuing another read/write operation whenever one completes until
completed, EOF if reading, or an error occurs.

Note that this MAX_RW_COUNT also applies to readv and writev, which
means we would need a similar mechanism there to handle this. This is
to be done in a follow-up.

Task-number: QTBUG-136763
Change-Id: I9bcb75587ae5e84cb80ea3950a569f60c4906617
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>