Specify TCP Keep Alive parameters via QNetworkRequest
Let the user define the TCP Keep Alive parameters via
QNetworkRequest.

The default values used by QNetworkAccessManager are defined in
QHttpNetworkConnectionChannel and can be overwritten by
environment variables.

These values can be also controled from the QNetworkRequest API.
These have the highest priority. If
nothing is defined, QNAM is going to use the default values. If the
environmental variables are defined, QNAM is going to use those. If
there are values provided via QNetworkRequest, QNAM is going to prefer
those.

[ChangeLog][QtNetwork][QNetworkRequest] Added new methods to specify
and get the current TCP KeepAlive parameters for the request.

Task-number: QTBUG-136625
Change-Id: Iafc485eb7b85214500d7c9205db1ecef67dc4b8c
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Marc Mutz <marc.mutz@qt.io>