Partially revert "qnumeric.h: add support for C23/C++26 <stdckdint.h>"
This reverts commit 08ebe3465cc2fce98662b5833b75503490f66265.

Reason for revert: the ckd_* functions are not constexpr. If we're going
to have to keep our implementations anyway, then we may as well skip
using the C23/C++26 feature. Maybe in C++29 they become constexpr.

I've left behind the requirement to not use plain char, in case we want
to switch to these later. That retains the old commit's changelog.

Fixes: QTBUG-142028
Change-Id: I6db786946ae31be16891840b03e5120eb8915d52
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>