tst_QSslKey: do not use EVP_PKEY_cmp on ML keys
We do not setup pkey for QSsl::MlDsa, and as a result EVP_PKEY_cmp, starting
from OpenSSL 3.5 is failing (retuns -2), essentially trying to compare
an empty EVP_PKEY with something we read from a pem file.
Previously we did not test those mldsa keys, thus the problem went unnoticed.

Change-Id: Icfa8ab7b703e1f5ae10d91bcaf37f5772b7eacdd
Pick-to: 6.10
Fixes: QTBUG-141723
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>