cmake: don't error out on deprecated macros
We exclude -Wdeprecated-declarations from -Werror builds, so we should
also exclude deprecated macros (#pragma clang deprecated(MACRO_NAME)),
soon to be used to deprecate Q_FLAGS/Q_ENUMS, as well.

Do so. This feature (and therefore the -Wno-error flag for it) was
introduced in LLVM Clang 14, which also aligns with AppleClang version
14 in this specific instance, so we can use a single version check.

Task-number: QTBUG-99060
Pick-to: 6.11 6.10 6.8 6.5
Change-Id: I9d4d1af5923b8f9523faf8f8415c168b19959652
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>