forkfd: configure our signal handler to restart the syscall
This should help solve problems with other libraries that don't have
proper EINTR_LOOP around their operations. Like most of <stdio.h> and
thus <iostreams>.

This change does not affect FreeBSD or Linux, where we don't install a
signal handler in the first place. And obviously, it will have no effect
if something else overrides our signal handler without this flag. If
that happens, please complain to the piece of software that forgot the
flag.

This test has also revealed that Windows presents a similar wake up: if
the grandchild process shares the child's stdin (which happens by
default for startDetached()) the child gets woken up from
WaitForSingleObject() when the grandchild dies and thus fgets() returns
error. That has nothing to do with QProcess, and AFAICT is simply the
way pipes work on Windows.

Fixes: QTBUG-142041
Pick-to: 6.10 6.8 6.5
Change-Id: I5900f238ec4a85ed49affffd29fc4248cc272929
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>