CMake: Remove faulty version numbers from qt_find_package(XCB) calls
The version in qt_find_package(XCB 0.1.1) is the version of libxcb, not
the version of one of its components. The version checks have been
ported from qmake but where never effective, because they check the
wrong thing and are so low that they're never failing.

The right version check for e.g.
    qt_find_package(XCB 0.3.8 MODULE COMPONENTS UTIL)
would be
    qt_find_package(XCB_UTIL 0.3.8 MODULE)
but that would change the XCB target names, and we evidently never ran
into compatibility problems in Qt 6 times, so let's remove the versions.

Change-Id: I42e8882a486c0184f96d86d73cd229bdb7c73057
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>