QAtomicScopedValueRollback: add some missing move()s
Unlike the non-atomic QScopedValueRollback (which has a moveOnly()
test to catch such mistakes), QAtomicScopedValueRollback was missing
some std::move()s in the implementation.

Add them.

We can't write a QScopedValueRollback-style moveOnly() test, because
atomic<> rquires the type to be copyable (what should load() return,
otherwise?), so the sharedPtr() test recently added is the closest we
can get.

Pick-to: 6.11 6.10 6.8
Change-Id: I6e0c05e34189a22b1b79a75ce73dc7c7c7e719c5
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>